requirements.txt
gym
-e git://github.com/atremblay/gym-cribbage.git@5307a2fbf7f0cb70a09aca10a56b0e0fb4ef7296#egg=gym-cribbage
torchvision
pyyaml